TS-3316: various trivial build fixes for 32bit systems.
Conflicts:
	CHANGES
	iocore/hostdb/HostDB.cc

+++ plugins/experimental/ts_lua/ts_lua_http_intercept.c
@@ -207,7 +207,7 @@ ts_lua_http_intercept_handler(TSCont contp, TSEvent event, void *edata)
 
   } else {
     mtxp = ictx->mctx->mutexp;
-    n = (int64_t) edata & 0xFFFF;
+    n = (int)((intptr_t) edata & 0xFFFF);
     TSMutexLock(mtxp);
     ret = ts_lua_http_intercept_run_coroutine(ictx, n);
   }
